requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
hiredis==2.3.2
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
rarfile==4.2
redis==5.0.4
sqlparse==0.5.0
typing_extensions==4.11.0
